+++ app/src/main/java/com/danosoftware/galaxyforce/sprites/game/starfield/StarTuple.java
@@ -1,26 +1,25 @@
 package com.danosoftware.galaxyforce.sprites.game.starfield;
 
 /**
- * Tuple of Star with it's initial Y position.
- * Helps computation and movement to a new Y position.
+ * Tuple of Star with it's initial Y position. Helps computation and movement to a new Y position.
  */
 class StarTuple {
 
-    private final Star star;
-    private final int initialY;
+  private final Star star;
+  private final float initialY;
 
-    StarTuple(
-            Star star,
-            int initialY) {
-        this.star = star;
-        this.initialY = initialY;
-    }
+  StarTuple(
+      Star star,
+      float initialY) {
+    this.star = star;
+    this.initialY = initialY;
+  }
 
-    Star getStar() {
-        return star;
-    }
+  Star getStar() {
+    return star;
+  }
 
-    int getInitialY() {
-        return initialY;
-    }
+  float getInitialY() {
+    return initialY;
+  }
 }
